Fixes #620 - bad_record_mac errors on SSL certificate negotiation
Fixes a minor problem with OpenSSL 0.9.8a which was causing the unit tests to
fail on Dapper (flight CD 3) - this should not be noticeable to most users.

Apparently the OP_ALL flag has been broken, at least 0.9.8a servers connecting
to 0.9.8a clients; other configurations are supported.  Clues imply that this
change has something to do with it:

http://www.openssl.org/news/secadv_20051011.txt

Not only was the flag set by default, it wasn't even being respected if it was
explicitly _un_set.  This fixes it so that the input argument is respected and
the default behavior is to turn it off.

The broken clients that are supported by this option are VERY old and VERY
broken.  The main code sslverify is designed to support, to wit, Q2Q, will not
even work over earlier versions of SSL - TLS is absolutely required, and there
is no non-OpenSSL implementation of this protocol, so leaving this flag off by
default makes sense.