Integrate Vertex with twisted.cred.
Author: glyph
Reviewer: radix

Fixes #693

This is only phase 1 of integration, as radix noted in his review; it simply
removes the ad-hoc password database interface and replaces it with a cred one.
q2qstandalone's functionality is unchanged, and the format is mostly the same -
however, passwords are now simply hex-encoded rather than md5'd due to cred's
constraints on hash-on-disk password databases.  This will break existing
vertex deployments, but who the heck has an existing vertex deployment?

There are 2 more phases to be done: phase 2) explicit integration with
Mantissa, including an IQ2QUser powerup which gives a user their Q2Q address,
and phase 3) hooks for responding to "INBOUND" and "LISTEN" messages from the
IQ2QUser avatar rather than internal to the service.