Initial kernel for development of Prime functionality.
This is a plugin architecture for adding things to a Divmod-themed GNOME tray
icon.  It refactors the only-half-functional vertex desktop client into a
plugin for that icon, and also provides a somewhat-useful Python console plugin
as well.

There's no Windows support, no tests, and no documentation.  This is entirely
proof-of-concept at this point.  Still, it might be fun to hack some stuff onto
(and now we can put Raygun, moe's GTK+ chat client framework, into the
repository somewhere as a plugin for this).

Author: glyph

Reviewer: washort

Fixes #943