requestCertificateForAddress should use one secure connection.
This eliminates a race condition by returning a signing response only
after a private certificate has been added to the local store.  The
certificate isn't added to the local store until after checkPassword's
Deferred resolves.  The code before this commit began a new secure
connection that didn't wait for that resolution.  Before the addition
of txscrypt, the credential checker returned a fired deferred, so not
waiting for checkPassword's Deferred to fire was the same as waiting
for it!